requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
defusedxml==0.7.1
Django==6.0.4
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
et_xmlfile==2.0.0
idna==3.11
numpy==2.4.4
oauthlib==3.3.1
openpyxl==3.1.5
pandas==3.0.2
pillow==12.2.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.33.1
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
urllib3==2.6.3
